Callback for keys that are not handled by the image view
Added a handler to init.js that is called when an unhandled key is pressed.

Handle a few extra image layer combinations: Shift + 1, 2 and 3 to show
some net combinations that are more interesting or useful to see.

You can show the values of (code,ctrl) if you hold down Alt while pressing
your key combination.