Autoscroll waveform view based on the cursor location
When a simulation run stops, check the (primary) cursor against the new sim
cycle and scroll the view if the cursor was positioned at the last cycle.
This will work for short runs of 1 and 2 half-cycles, for which we want to
observe the signal traces step by step anyways. All other runs do not scroll
the view.